Dump commit - NOT WORKING
This is just to make sure I save state. Jason and I ended up going through a lot of variations attempting to get this code to work. The result is that many parts of the code are in flux.

A few interesting points.

Extending a class with `Thread` causes it to have very strange variable behavior. In this case, it meant that code like

$queue->enqueue("my thing")
$queue->dequeue()

would cause an exception because there was nothing in the queue. Yep. Weird.

The code in tmp.php represents the closest thing to a "correct" version of using pthreads to do what we want. It's a matter of instantiating a pool with the correct worker class and correct worker class parameters.

We've stopped for now because we found out that the maintainer of pthreads is against using pthreads in a web environment and will be adding code to actively block the use of pthreads in those situations.

More discussion to follow.